Teach the Emacs mode to indent multi-line strings
When indenting a region, if in a multi-line string,
don't update the state, since we're still in a string.

Previously, typing C-M-\ or M-x indent-region RET on the following:

    somefunction() ->
        S = <<"
    text1 somemore1
    text2 somemore2
    ">>,
        io:put_chars(S).

resulted in this (note the "text2" line):

    somefunction() ->
        S = <<"
    text1 somemore1
              text2 somemore2
    ">>,
        io:put_chars(S).

Now, the indentation inside the multi-line string is left unchanged.
This is in line with how the Emacs C and Lisp mode treats multi-line
strings when indenting regions.